Add Indexed FlowStepper as property of FlowStepper (FlowStepper.Indexed)
