+++ packages/react/src/components/navigation/FlowStepper/index.tsx
@@ -170,3 +170,55 @@ function Step({ children }: StepProps) {
 
 FlowStepper.Step = Step;
 export default FlowStepper;
+
+export type IndexedStepProps = StepProps & {
+  /**
+   * String to identify the step. Must be different from sibling steps's `key` prop.
+   */
+  itemKey: string;
+};
+
+type IndexedStepperChild = React.ReactElement<IndexedStepProps>;
+
+export type IndexedProps<ExtraProps> = Omit<Props<ExtraProps>, "activeIndex" | "children"> & {
+  /**
+   * The key of the active step
+   */
+  activeKey: string;
+
+  /**
+   * A list of children representing each step of the flow.
+   */
+  children: IndexedStepperChild | IndexedStepperChild[];
+};
+
+/**
+ * This is a FlowStepper where each children must have an `itemKey: string` prop
+ * and the active step is defined by the `activeKey: string` prop.
+ *
+ * This allows for usages with a lot of steps where dealing with indices could be
+ * painful and error prone (for instance inserting/removing a step somewhere would shift
+ * the indices of the following steps and navigation would be impacted).
+ *
+ * By using string identifiers (`itemKey`) for each step, it's more "human readable"
+ * and less error prone to setup a navigation logic between steps.
+ */
+function FlowStepperIndexed<ExtraProps>(props: IndexedProps<ExtraProps>) {
+  const { activeKey, children, ...otherProps } = props;
+  const activeIndex = React.Children.toArray(children).findIndex((child) => {
+    const res = isElement(child) && child.props.itemKey === activeKey;
+    return res;
+  });
+  return (
+    <FlowStepper {...otherProps} activeIndex={activeIndex}>
+      {children}
+    </FlowStepper>
+  );
+}
+
+function IndexedStep({ children }: IndexedStepProps) {
+  return <>{children}</>;
+}
+
+FlowStepperIndexed.Step = IndexedStep;
+FlowStepper.Indexed = FlowStepperIndexed;
